A JQuery Widget
Technologies Used- JSON,J Query,AJAX,JavaScript.

Provide five circles for the user to select or click on.
When the user hovers their mouse over a circle, that circle and all
circles to the left must be filled with a color, such as yellow.
Provide a configurable number of circles for the user to click on.Allow
the number of circles to be updated by a text-box.
This is to simulate an administration widget.we will not be adding any
form of security to the widget.
Final feature to my rating widget,the ability to send the rating to a
service.
The service has already been created. The service will act as a simple
echo service, taking the rating you submit inside a JSON object, and
then sending back a JSON containing a message